+++ src/platform/roblox/RobloxColorProvider.cpp
@@ -2,6 +2,23 @@
 
 #include "LSP/ColorProvider.hpp"
 
+static std::string formatDouble(double d)
+{
+    std::string s = std::to_string(d);
+
+    // Strip trailing decimals that are redundant
+    size_t decimalPos = s.find('.');
+    if (decimalPos != std::string::npos)
+    {
+        size_t lastNonZero = std::max(s.find_last_not_of('0'), decimalPos);
+        if (lastNonZero != std::string::npos)
+            s.erase(lastNonZero + 1);
+        if (s.back() == '.')
+            s.pop_back();
+    }
+    return s;
+}
+
 struct RobloxColorVisitor : public Luau::AstVisitor
 {
     const TextDocument* textDocument;
@@ -128,8 +145,8 @@ lsp::ColorPresentationResult RobloxPlatform::colorPresentation(const lsp::ColorP
     lsp::ColorPresentationResult presentations;
 
     // Add Color3.new
-    presentations.emplace_back(lsp::ColorPresentation{"Color3.new(" + std::to_string(params.color.red) + ", " + std::to_string(params.color.green) +
-                                                      ", " + std::to_string(params.color.blue) + ")"});
+    presentations.emplace_back(lsp::ColorPresentation{
+        "Color3.new(" + formatDouble(params.color.red) + ", " + formatDouble(params.color.green) + ", " + formatDouble(params.color.blue) + ")"});
 
     // Convert to RGB values
     RGB rgb{
@@ -145,7 +162,7 @@ lsp::ColorPresentationResult RobloxPlatform::colorPresentation(const lsp::ColorP
     // Add Color3.fromHSV
     HSV hsv = rgbToHsv(rgb);
     presentations.emplace_back(
-        lsp::ColorPresentation{"Color3.fromHSV(" + std::to_string(hsv.h) + ", " + std::to_string(hsv.s) + ", " + std::to_string(hsv.v) + ")"});
+        lsp::ColorPresentation{"Color3.fromHSV(" + formatDouble(hsv.h) + ", " + formatDouble(hsv.s) + ", " + formatDouble(hsv.v) + ")"});
 
     // Add Color3.fromHex
     presentations.emplace_back(lsp::ColorPresentation{"Color3.fromHex(\"" + rgbToHex(rgb) + "\")"});